Added option --cleanenv to singularity
The --cleanenv option passed to singularity is necessary to run Kraken2 because of how CHPC is setup: the version of Perl on CHPC causes issues when running the container with singularity because by default Singularity passes all env variables to the container, which included paths to Perl libraries specified on CHPC. This resulted in the container trying to access incompatible libraries. By using the --cleanenv option, no env variable is passed to the container.